WIP: crc32Combine function corrected + fuzzed.
Thanks to some assistance from Github Copilot, I was able to discern
that I needed to do some additional transformation of the front-side
checksum before XOR'ing with the back-side checksum, so that the final
checksum would be equivalent to:

  crc32(AB) = crc32(A0) + crc32(B)

This required some extra XORs with 0xffffffff, which I found surprising.
The fuzz tests pass now, which is a very promising sign.

Signed-off-by: Philip Conrad <philip@chariot-chaser.net>